Separate out a tee client.
This allows the pcapng pipe to be consistent, while not blocking
on the main teeing process.

Before this client-server architecture, the pcapng pipe would get full
and then block the proxying. Switching the pcap file open to O_NONBLOCK
doesn't work eiether because that breaks the consistency of the pcapng
format.